move library modules to new namespace
