+++ demo/android/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         tools:targetApi="31"
         >
         <activity
-            android:name=".MainActivity"
+            android:name="com.dragselectcompose.demo.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.DragSelectCompose"
             >
