+++ demo/kmm/shared/src/commonMain/kotlin/dev/jordond/dragselectcompose/demo/App.kt
@@ -15,7 +15,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import dev.jordond.dragselectcompose.rememberDragSelectState
+import com.dragselectcompose.core.rememberDragSelectState
 
 @Composable
 fun App() {
