+++ dragselectcompose/src/commonMain/kotlin/com/dragselectcompose/Placeholder.kt
@@ -1,0 +1,9 @@
+package com.dragselectcompose
+
+/**
+ * `:dragselectcompose` includes the `:core` and :extensions` libraries.
+ *
+ * This library is a container for both of them.
+ */
+@Suppress("unused")
+internal object Placeholder
