--- dragselectcompose/src/commonMain/kotlin/dev/jordond/dragselectcompose/Placeholder.kt
@@ -1,8 +1,0 @@
-package dev.jordond.dragselectcompose
-
-/**
- * `:dragselectcompose` includes the `:core` and :extensions` libraries.
- *
- * This library is a container for both of them.
- */
-internal object Placeholder
